# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис управления складами и товарами.",
        "title": "Go-Freight Inventory API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/warehouses": {
            "get": {
                "description": "Возвращает список всех зарегистрированных складов в системе",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "warehouses"
                ],
                "summary": "Получить список всех складов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_kelar1s_go-freight_internal_inventory_transport_rest_dto.WarehouseResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelar1s_go-freight_internal_inventory_transport_rest_dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Регистрирует новый склад с указанием названия и адреса",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "warehouses"
                ],
                "summary": "Создать новый склад",
                "parameters": [
                    {
                        "description": "Данные склада",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_kelar1s_go-freight_internal_inventory_transport_rest_dto.CreateWarehouseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelar1s_go-freight_internal_inventory_transport_rest_dto.WarehouseResponse"
                        }
                    },
                    "400": {
                        "description": "Отсутствует название или локация",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelar1s_go-freight_internal_inventory_transport_rest_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelar1s_go-freight_internal_inventory_transport_rest_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/warehouses/{id}": {
            "get": {
                "description": "Возвращает детальную информацию о складе по его идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "warehouses"
                ],
                "summary": "Получить склад по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID склада",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelar1s_go-freight_internal_inventory_transport_rest_dto.WarehouseResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelar1s_go-freight_internal_inventory_transport_rest_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Склад не найден",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelar1s_go-freight_internal_inventory_transport_rest_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelar1s_go-freight_internal_inventory_transport_rest_dto.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Полностью обновляет название и локацию склада по его ID",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "warehouses"
                ],
                "summary": "Обновить данные склада",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID склада",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные склада",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_kelar1s_go-freight_internal_inventory_transport_rest_dto.UpdateWarehouseRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Неверный формат ID или неполные данные",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelar1s_go-freight_internal_inventory_transport_rest_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Склад не найден",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelar1s_go-freight_internal_inventory_transport_rest_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelar1s_go-freight_internal_inventory_transport_rest_dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет склад по его ID",
                "tags": [
                    "warehouses"
                ],
                "summary": "Удалить склад",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID склада",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelar1s_go-freight_internal_inventory_transport_rest_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelar1s_go-freight_internal_inventory_transport_rest_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_kelar1s_go-freight_internal_inventory_transport_rest_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_kelar1s_go-freight_internal_inventory_transport_rest_dto.CreateWarehouseRequest": {
            "type": "object",
            "properties": {
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_kelar1s_go-freight_internal_inventory_transport_rest_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_kelar1s_go-freight_internal_inventory_transport_rest_dto.UpdateWarehouseRequest": {
            "type": "object",
            "properties": {
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_kelar1s_go-freight_internal_inventory_transport_rest_dto.WarehouseResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}